code/requirements.txt
accelerate==0.30.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm @ file:///croot/comm_1671231121260/work
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.8
evaluate==0.4.2
exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.14.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.2
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imagecodecs==2024.6.1
imageio==2.34.1
importlib-metadata @ file:///croot/importlib-metadata_1678997070253/work
importlib_resources==6.4.0
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1694181358621/work
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2==3.1.4
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
networkx==3.2.1
notebook==7.2.0
notebook_shim==0.2.4
numcodecs==0.12.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.2
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
pillow @ file:///croot/pillow_1714398848491/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
protobuf==4.25.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydeck==0.9.1
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
referencing==0.35.1
regex==2024.5.10
requests @ file:///croot/requests_1707355572290/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-image==0.22.0
scipy==1.13.1
Send2Trash==1.8.3
shapely==2.0.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
stqdm==0.0.5
streamlit==1.35.0
tabulate==0.9.0
tenacity==8.3.0
terminado==0.18.1
tifffile==2024.5.22
tiffslide==2.4.0
timm==1.0.7
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==1.13.0
torch-summary==1.4.5
torchaudio==0.13.0
torchinfo==1.8.0
torchvision==0.14.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1700840488771/work
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
wsinfer==0.6.1
wsinfer-zoo==0.6.2
xxhash==3.4.1
yarl==1.9.4
zarr==2.18.2
zipp @ file:///croot/zipp_1672387121353/work
